backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
pillow
django-cleanup
google-generativeai
psycopg2-binary